requirements.txt
astroid==1.6.0
certifi==2017.11.5
chardet==3.0.4
coverage==4.4.2
defusedxml==0.5.0
Django==2.0.1
django-allauth==0.34.0
django-cors-headers==2.1.0
django-nose==1.4.5
django-rest-auth==0.9.2
django-rest-framework==0.1.0
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
google-api-python-client==1.6.5
httplib2==0.10.3
idna==2.6
isort==4.2.15
lazy-object-proxy==1.3.1
mccabe==0.6.1
nose==1.3.7
oauth2client==4.1.2
oauthlib==2.0.6
pep8==1.7.1
Pillow==5.0.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
PyJWT==1.5.3
pylint==1.8.1
python3-openid==3.1.0
pytz==2017.3
requests==2.18.4
requests-oauthlib==0.8.0
rsa==3.4.2
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.6.0
uritemplate==3.0.0
urllib3==1.22
wrapt==1.10.11
